third_party/WebKit/LayoutTests/inspector/styles/variables/css-variables.html

<html>
<head>
<style>
#inspected {
    var-a: green;
    color: var(a);
}

</style>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    WebInspector.showPanel("elements");
    InspectorTest.selectNodeAndWaitForStylesWithComputed("inspected", dumpAllStyles);

    function dumpAllStyles()
    {
        InspectorTest.dumpSelectedElementStyles();
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that webkit css variables can be loaded correctly.
</p>

<div id="inspected">Text</div>

</body>
</html>
