third_party/WebKit/LayoutTests/inspector/styles/variables/css-variables-expected.txt

Tests that webkit css variables can be loaded correctly.

Text
[expanded] 
color: rgb(0, 128, 0);
    #inspected - var(a) css-variables.html:4
display: block;
    div - block user agent stylesheet

[expanded] 
element.style  { ()

[expanded] 
#inspected  { (css-variables.html:4 -> css-variables.html:4:1)
var-a: green;
color: var(a);

[expanded] 
div  { (user agent stylesheet)
display: block;
